Added a GUI
This commit introduces a graphical user interface for the project. The game
offers the player buttons to select and compares the player selection with the
random computer choice. The result of every round is shown until either
computer or player reaches the score of 5 then the game ends and the player is
required to reload the page.